src/requirements.txt
confluent-kafka==1.6.1
fastapi==0.63.0
pydantic==1.8.1
uvicorn==0.13.4
pytest==6.2.3
pytest_schema==0.1.0
requests==2.25.1
cassandra-driver==3.25.0